+++ apps/web_ivi/static/index.html
@@ -5,8 +5,9 @@
     <title>Demo Web IVI</title>
     <!-- include styles.css -->
     <link rel="stylesheet" type="text/css" href="static/styles.css">
-    <script type="text/javascript" src="static/vehicle-dynamics.js"></script>
-    <script src="static/turn-signals.js"></script>
+    <!-- <script type="text/javascript" src="static/vehicle-dynamics.js"></script>
+    <script src="static/turn-signals.js"></script> -->
+    <script type="module" src="static/ecal.js"></script>
 </head>
 <body>
     <div class="row-main">
