+++ apps/web_ivi/static/vehicle-dynamics.js
@@ -22,7 +22,7 @@ const speedTorqueRelation = [{
 }];
 
 /* update the speed-value */
-function updateSpeedAndTorque(speed_km_h) {
+export function updateSpeedAndTorque(speed_km_h) {
     // update the speed value fluently by a number
     let speedValue = document.getElementById("speed-value");
     let torqueValue = document.getElementById("torque-value");
@@ -43,27 +43,4 @@ function updateSpeedAndTorque(speed_km_h) {
     }
 
     speedValue.textContent = speed_km_h;
-}
-
-// read from server-side event
-const eventSource = new EventSource("/vehicle-dynamics");
-
-eventSource.onopen = function (_event) {
-    console.log("Connection to /vehicle-dynamics opened");
-}
-
-// add event listener for the event
-eventSource.addEventListener("vehicle-dynamics", function (event) {
-    let raw_data = event.data;
-    let vehicle_dynamics = JSON.parse(raw_data);
-    console.log(vehicle_dynamics);
-    let speed_km_h = parseInt(parseFloat(vehicle_dynamics.signals.speedDisplayed) * 3.6)
-    console.log("Received speed_km_h: ", speed_km_h);
-    updateSpeedAndTorque(speed_km_h);
-});
-
-// close the connection on error
-eventSource.onerror = function (event) {
-    console.log("Error: " + event);
-    eventSource.close();
 }
